perf(Perfomance): :zap: Rewrite All Nav Component
Rewrite all component in Navigation category, so it's look more clean
