+++ src/components/elements/Image/type.ts
@@ -3,6 +3,6 @@ export type ImageType = {
   alt: string;
   type?: "avatar" | "rounded";
   className?: string;
-  width: "xxs" | "xs" | "sm" | "md" | "lg" | "xl" | "xxl";
-  height: "xxs" | "xs" | "sm" | "md" | "lg" | "xl" | "xxl";
+  width: number;
+  height: number;
 };
