+++ src/components/layouts/Header/index.tsx
@@ -2,8 +2,8 @@ import { SignedIn, SignedOut, UserButton } from "@clerk/nextjs";
 import Image from "next/image";
 import Link from "next/link";
 import { Button } from "../../ui/button";
-import NavItems from "../../navigations/Menubar/NavItems";
-import MobileNav from "../../navigations/Menubar/MobileNav";
+import NavItems from "../../navigations/Menubar";
+import MobileNav from "../../navigations/MobileNav";
 
 const Header = () => {
   return (
