+++ src/components/navigations/Link/index.tsx
@@ -1,11 +1,9 @@
+import type { LinkType } from "./type";
+import type { ChildType } from "@/types";
 import Link from "next/link";
 
-const index = ({
-  children,
-  href,
-}: {
-  children: React.ReactNode;
-  href: string;
-}) => <Link href="#events">{children}</Link>;
+const index = ({ children, href }: LinkType & ChildType) => (
+  <Link href={href}>{children}</Link>
+);
 
 export default index;
