+++ src/components/navigations/Link/type.ts
@@ -1,0 +1,3 @@
+export type LinkType = {
+  href: string;
+};
