+++ src/components/navigations/index.ts
@@ -1,5 +1,5 @@
 export { default as Link } from "./Link";
-// export { default as Menubar } from "./Menubar";
+export { default as Menubar } from "./Menubar";
 // export { default as NavItems } from "./Menubar/NavItems";
 // export { default as MobileMenu } from "./MobileMenu";
 // export { default as MobileNavItems } from "./MobileMenu/MobileNavItems"
