+++ src/utils/getPathName.ts
@@ -1,0 +1,4 @@
+"use client";
+import { usePathname } from "next/navigation";
+
+export const GetPathName = () => usePathname();
